refactor(Date): nowモジュールの名前をnowSimpleに変更
