+++ package/main/src/Simple/Date/index.ts
@@ -1,1 +1,1 @@
-export * from "./now";
+export * from "./nowSimple";
